+++ src/api/style/VerticalAlignment.ts
@@ -1,0 +1,7 @@
+export enum VerticalAlignment {
+  Default = '',
+  Top = 'top',
+  Middle = 'middle',
+  Bottom = 'bottom',
+  Baseline = 'baseline'
+}