+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **style:** add common styles
 - **style:** allow page break to be before or after a paragraph, closes [#31](https://github.com/connium/simple-odf/issues/31)
 - **style:** add keep-with-next flag
+- **style:** add orphan & widdow control, paragraph background, line height
 - **docs:** add a realistic example
 
 ### Changed